+++ src/Parser/readEnum.ts
@@ -120,7 +120,7 @@ export function readEnum(
     }
     const croppedLine = line.slice(i);
     let iterMatch = croppedLine.match(
-      /^\s*(?:\w+\s*:\s*)?([A-Za-z_]+\w*)(?:\s*\=.+?(?=(?:\,|\/\*|\n|\/\/)))?/
+      /^\s*(?:\w+\s*:\s*)?([A-Za-z_]+\w*)(?:\s*\=.+?(?=(?:\,|\/\*|$|\/\/)))?/
     );
     if (!iterMatch || isBlockComment) {
       i++;